refactor(lanes): consolidate prelude onto structural, remove legacy basic prelude
