+++ demo/community_call_jan_28_2026.py
@@ -23,8 +23,6 @@ def show_lanes(arch: ArchSpec):
 logical_arch = logical.get_arch_spec()
 show_lanes(logical_arch)
 
-# %%
-from bloqade.lanes._prelude import kernel  # noqa F402
 from bloqade.lanes.bytecode.encoding import (  # noqa F402
     Direction,
     LocationAddress,
@@ -33,6 +31,9 @@ from bloqade.lanes.bytecode.encoding import (  # noqa F402
     ZoneAddress,
 )
 from bloqade.lanes.dialects import move  # noqa F402
+
+# %%
+from bloqade.lanes.prelude import kernel  # noqa F402
 
 
 @kernel
