+++ python/tests/rewrite/move2squin/test_gate.py
@@ -7,11 +7,11 @@ from kirin.analysis import forward
 from kirin.dialects import func, ilist, py
 
 from bloqade import qubit, squin, types as bloqade_types
-from bloqade.lanes._prelude import kernel
 from bloqade.lanes.analysis import atom
 from bloqade.lanes.arch.gemini.logical import get_arch_spec
 from bloqade.lanes.bytecode.encoding import LocationAddress, ZoneAddress
 from bloqade.lanes.dialects import move
+from bloqade.lanes.prelude import kernel
 from bloqade.lanes.rewrite.move2squin import base, gates
