+++ python/tests/validation/test_address.py
@@ -3,7 +3,6 @@ from kirin import ir
 from kirin.ir.exception import ValidationErrorGroup
 from kirin.validation import ValidationSuite
 
-from bloqade.lanes._prelude import kernel as lanes_kernel
 from bloqade.lanes.arch.gemini.logical import validation
 from bloqade.lanes.bytecode.encoding import (
     Direction,
@@ -12,6 +11,7 @@ from bloqade.lanes.bytecode.encoding import (
     ZoneAddress,
 )
 from bloqade.lanes.dialects import move
+from bloqade.lanes.prelude import kernel as lanes_kernel
 from bloqade.lanes.types import State
